Multiple persistent areas, tweaked defence system

+++ Roguelike/EffectHealthRegeneration.cpp
@@ -8,6 +8,6 @@ std::string EffectHealthRegeneration::getDescription(){
 }
 
 void EffectHealthRegeneration::apply(Creature &creature){
-	creature.health += creature.healthMax * amount;
+	creature.health += (int)(creature.healthMax * amount);
 	if (creature.health > creature.healthMax) creature.health = creature.healthMax;
 }
